[Exchanges] Refactor architecture
Add connectors
