+++ octobot_trading/exchanges/__init__.pxd
@@ -13,66 +13,54 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
-from octobot_trading.exchanges cimport exchanges
 from octobot_trading.exchanges.exchanges cimport (
     ExchangeConfiguration,
     Exchanges,
 )
 
-from octobot_trading.exchanges cimport exchange_config_data
 from octobot_trading.exchanges.exchange_config_data cimport (
     ExchangeConfig,
 )
 
-from octobot_trading.exchanges cimport traders
 from octobot_trading.exchanges.traders cimport (
     Trader,
     TraderSimulator,
 )
 
-from octobot_trading.exchanges cimport abstract_exchange
 from octobot_trading.exchanges.abstract_exchange cimport (
     AbstractExchange,
 )
 
-from octobot_trading.exchanges cimport websockets
 from octobot_trading.exchanges.websockets cimport (
     AbstractWebsocket,
-    OctoBotWebSocketClient,
-    force_disable_web_socket,
+force_disable_web_socket,
     check_web_socket_config,
     search_websocket_class,
 )
 
-from octobot_trading.exchanges cimport exchange_manager
 from octobot_trading.exchanges.exchange_manager cimport (
     ExchangeManager,
 )
-from octobot_trading.exchanges cimport exchange_util
-from octobot_trading.exchanges.exchange_util cimport (
+from octobot_trading.exchanges.util.exchange_util cimport (
     get_margin_exchange_class,
     get_future_exchange_class,
     get_spot_exchange_class,
     get_order_side,
 )
-from octobot_trading.exchanges cimport util
 from octobot_trading.exchanges.util cimport (
     ExchangeMarketStatusFixer,
     is_ms_valid,
 )
-from octobot_trading.exchanges cimport types
 from octobot_trading.exchanges.types cimport (
     FutureExchange,
     WebsocketExchange,
     MarginExchange,
     SpotExchange,
 )
-from octobot_trading.exchanges cimport implementations
 from octobot_trading.exchanges.implementations cimport (
     ExchangeSimulator,
     CCXTExchange,
 )
-from octobot_trading.exchanges cimport exchange_builder
 from octobot_trading.exchanges.exchange_builder cimport (
     ExchangeBuilder,
 )
