+++ octobot_trading/exchanges/abstract_websocket_exchange.pxd
