--- octobot_trading/exchanges/channel/__init__.py
@@ -1,30 +1,0 @@
-#  Drakkar-Software OctoBot-Trading
-#  Copyright (c) Drakkar-Software, All rights reserved.
-#
-#  This library is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 3.0 of the License, or (at your option) any later version.
-#
-#  This library is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-#  Lesser General Public License for more details.
-#
-#  You should have received a copy of the GNU Lesser General Public
-#  License along with this library.
-
-from octobot_trading.exchanges.channel import exchange_channels
-from octobot_trading.exchanges.channel.exchange_channels import (
-    requires_refresh_trigger,
-    create_exchange_channels,
-    create_exchange_producers,
-    create_authenticated_producer_from_parent,
-)
-
-__all__ = [
-    "requires_refresh_trigger",
-    "create_exchange_channels",
-    "create_exchange_producers",
-    "create_authenticated_producer_from_parent",
-]
