+++ octobot_trading/exchanges/connectors/ccxt_exchange.pxd
