+++ octobot_trading/exchanges/connectors/exchange_simulator.pxd
