+++ octobot_trading/exchanges/connectors/websocket_connector.pxd
