+++ octobot_trading/exchanges/util/websockets_util.py
