--- octobot_trading/exchanges/websockets/__init__.pxd
@@ -1,39 +1,0 @@
-#  Drakkar-Software OctoBot-Trading
-#  Copyright (c) Drakkar-Software, All rights reserved.
-#
-#  This library is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 3.0 of the License, or (at your option) any later version.
-#
-#  This library is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-#  Lesser General Public License for more details.
-#
-#  You should have received a copy of the GNU Lesser General Public
-#  License along with this library.
-
-from octobot_trading.exchanges.websockets cimport octobot_websocket
-from octobot_trading.exchanges.websockets cimport abstract_websocket
-from octobot_trading.exchanges.websockets cimport websockets_util
-
-from octobot_trading.exchanges.websockets.octobot_websocket cimport (
-    OctoBotWebSocketClient,
-)
-from octobot_trading.exchanges.websockets.abstract_websocket cimport (
-    AbstractWebsocket,
-)
-from octobot_trading.exchanges.websockets.websockets_util cimport (
-    force_disable_web_socket,
-    check_web_socket_config,
-    search_websocket_class,
-)
-
-__all__ = [
-    "OctoBotWebSocketClient",
-    "AbstractWebsocket",
-    "force_disable_web_socket",
-    "check_web_socket_config",
-    "search_websocket_class",
-]
